reservation menu and role base access but role base and isloggeed feature halka errror aaya ko xa:

+++ server/src/database/models/category.Model.ts
@@ -13,12 +13,9 @@ import {
   modelName: "Category",
 })
 class Category extends Model {
-  // id
   @PrimaryKey
   @AutoIncrement
-  @Column({
-    type: DataType.BIGINT,
-  })
+  @Column({ type: DataType.INTEGER })
   declare id: number;
   // categoryName
   @Column({
